fix: do not delete deprecated metrics on additional label change

+++ status/controller.go
@@ -203,15 +203,15 @@ func (c *Controller[T]) reconcile(ctx context.Context, req reconcile.Request, o 
 		if errors.IsNotFound(err) {
 			c.observedConditions.Delete(req)
 			c.observedGaugeLabels.Delete(req)
-			c.deletePartialMatchGaugeMetric(c.ConditionCount, ConditionCount, map[string]string{
+			c.deletePartialMatchGaugeMetric(c.ConditionCount, ConditionCount, true, map[string]string{
 				MetricLabelNamespace: req.Namespace,
 				MetricLabelName:      req.Name,
 			}, nil)
-			c.deletePartialMatchGaugeMetric(c.ConditionCurrentStatusSeconds, ConditionCurrentStatusSeconds, map[string]string{
+			c.deletePartialMatchGaugeMetric(c.ConditionCurrentStatusSeconds, ConditionCurrentStatusSeconds, true, map[string]string{
 				MetricLabelNamespace: req.Namespace,
 				MetricLabelName:      req.Name,
 			}, nil)
-			c.deletePartialMatchGaugeMetric(c.TerminationCurrentTimeSeconds, TerminationCurrentTimeSeconds, map[string]string{
+			c.deletePartialMatchGaugeMetric(c.TerminationCurrentTimeSeconds, TerminationCurrentTimeSeconds, true, map[string]string{
 				MetricLabelNamespace: req.Namespace,
 				MetricLabelName:      req.Name,
 			}, nil)
@@ -275,15 +275,17 @@ func (c *Controller[T]) reconcile(ctx context.Context, req reconcile.Request, o 
 	}
 
 	for _, observedCondition := range observedConditions.List() {
-		if currentCondition := currentConditions.Get(observedCondition.Type); currentCondition == nil || currentCondition.Status != observedCondition.Status || currentCondition.Reason != observedCondition.Reason || !maps.Equal(c.toAdditionalGaugeMetricLabels(o), observedGaugeLabels) {
-			c.deletePartialMatchGaugeMetric(c.ConditionCount, ConditionCount, map[string]string{
+		currentCondition := currentConditions.Get(observedCondition.Type)
+		conditionChanged := currentCondition == nil || currentCondition.Status != observedCondition.Status || currentCondition.Reason != observedCondition.Reason
+		if conditionChanged || !maps.Equal(c.toAdditionalGaugeMetricLabels(o), observedGaugeLabels) {
+			c.deletePartialMatchGaugeMetric(c.ConditionCount, ConditionCount, conditionChanged, map[string]string{
 				MetricLabelNamespace:       req.Namespace,
 				MetricLabelName:            req.Name,
 				pmetrics.LabelType:         observedCondition.Type,
 				MetricLabelConditionStatus: string(observedCondition.Status),
 				pmetrics.LabelReason:       observedCondition.Reason,
 			}, observedGaugeLabels)
-			c.deletePartialMatchGaugeMetric(c.ConditionCurrentStatusSeconds, ConditionCurrentStatusSeconds, map[string]string{
+			c.deletePartialMatchGaugeMetric(c.ConditionCurrentStatusSeconds, ConditionCurrentStatusSeconds, conditionChanged, map[string]string{
 				MetricLabelNamespace:       req.Namespace,
 				MetricLabelName:            req.Name,
 				pmetrics.LabelType:         observedCondition.Type,
@@ -356,9 +358,9 @@ func (c *Controller[T]) setGaugeMetric(current pmetrics.GaugeMetric, deprecated 
 	}
 }
 
-func (c *Controller[T]) deletePartialMatchGaugeMetric(current pmetrics.GaugeMetric, deprecated pmetrics.GaugeMetric, labels, additionalLabels map[string]string) {
+func (c *Controller[T]) deletePartialMatchGaugeMetric(current pmetrics.GaugeMetric, deprecated pmetrics.GaugeMetric, deleteDeprecatedMetrics bool, labels, additionalLabels map[string]string) {
 	current.DeletePartialMatch(lo.Assign(labels, additionalLabels))
-	if c.emitDeprecatedMetrics {
+	if c.emitDeprecatedMetrics && deleteDeprecatedMetrics {
 		labels[pmetrics.LabelKind] = c.gvk.Kind
 		labels[pmetrics.LabelGroup] = c.gvk.Group
 		deprecated.DeletePartialMatch(labels)
